Doc: add authordep in example
If the PluginBundle is used to build the distribution of the
PluginBundle itself (thanks to [Bootstrap::lib]), it is important to
have all its compile-time dependencies advertised as authordep in
dist.ini so `dzil authordep` can report them.

So the '; authordep Dist::Zilla::Role::BundleDeps' is added in the
example of usage for this use case.